refactor out EmbeddedComponentCoordinator
