+++ connect/src/main/java/com/stripe/android/connect/di/StripeConnectComponent.kt
@@ -1,30 +1,31 @@
 package com.stripe.android.connect.di
 
 import com.stripe.android.connect.BuildConfig
-import com.stripe.android.connect.EmbeddedComponentManager
-import com.stripe.android.connect.PrivateBetaConnectSDK
 import com.stripe.android.connect.analytics.ConnectAnalyticsModule
 import com.stripe.android.connect.analytics.ConnectAnalyticsServiceFactory
+import com.stripe.android.connect.manager.EmbeddedComponentManagerComponent
 import com.stripe.android.core.injection.CoreCommonModule
 import com.stripe.android.core.injection.ENABLE_LOGGING
 import dagger.BindsInstance
 import dagger.Component
+import dagger.Module
 import javax.inject.Named
+import javax.inject.Provider
 import javax.inject.Singleton
 
 @Singleton
 @Component(
     modules = [
+        StripeConnectComponentModule::class,
         CoreCommonModule::class,
         ConnectAnalyticsModule::class,
     ]
 )
-@OptIn(PrivateBetaConnectSDK::class)
 internal interface StripeConnectComponent {
 
     val analyticsServiceFactory: ConnectAnalyticsServiceFactory
 
-    fun inject(embeddedComponentManager: EmbeddedComponentManager)
+    val coordinatorComponentProvider: Provider<EmbeddedComponentManagerComponent.Factory>
 
     @Component.Builder
     interface Builder {
@@ -41,3 +42,6 @@ internal interface StripeConnectComponent {
                 .build()
     }
 }
+
+@Module(subcomponents = [EmbeddedComponentManagerComponent::class])
+internal class StripeConnectComponentModule
