+++ connect/src/main/java/com/stripe/android/connect/manager/EmbeddedComponentManagerScope.kt
@@ -1,0 +1,7 @@
+package com.stripe.android.connect.manager
+
+import javax.inject.Scope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+internal annotation class EmbeddedComponentManagerScope
