+++ connect/src/main/java/com/stripe/android/connect/webview/StripeConnectWebViewContainerViewModel.kt
@@ -92,6 +92,8 @@ internal class StripeConnectWebViewContainerViewModel(
     private val loggerTag = javaClass.simpleName
     private val _stateFlow = MutableStateFlow(StripeConnectWebViewContainerState())
 
+    private val componentCoordinator get() = embeddedComponentManager.coordinator
+
     /**
      * Flow of the container state.
      */
@@ -125,7 +127,7 @@ internal class StripeConnectWebViewContainerViewModel(
             // Load local static html for instrumentation tests.
             webView.loadData(getTestHtml(), "text/html", "utf-8")
         } else {
-            webView.loadUrl(embeddedComponentManager.getStripeURL(embeddedComponent))
+            webView.loadUrl(componentCoordinator.getStripeURL(embeddedComponent))
         }
 
         analyticsService.track(ConnectAnalyticsEvent.ComponentViewed(stateFlow.value.pageViewId))
@@ -136,7 +138,7 @@ internal class StripeConnectWebViewContainerViewModel(
             owner.repeatOnLifecycle(Lifecycle.State.STARTED) {
                 launch {
                     // Bind appearance changes in the manager to the WebView (only when page is loaded).
-                    embeddedComponentManager.appearanceFlow
+                    componentCoordinator.appearanceFlow
                         .collectLatest { appearance ->
                             updateState { copy(appearance = appearance) }
                             if (stateFlow.value.pageViewId != null) {
@@ -161,7 +163,7 @@ internal class StripeConnectWebViewContainerViewModel(
             updateState { copy(isNativeLoadingIndicatorVisible = !receivedSetOnLoaderStart) }
 
             val pageLoadUrl = Uri.parse(url)
-            val expectedUrl = Uri.parse(embeddedComponentManager.getStripeURL(embeddedComponent))
+            val expectedUrl = Uri.parse(componentCoordinator.getStripeURL(embeddedComponent))
             if (
                 pageLoadUrl.scheme != expectedUrl.scheme ||
                 pageLoadUrl.host != expectedUrl.host ||
@@ -257,11 +259,11 @@ internal class StripeConnectWebViewContainerViewModel(
         }
 
         override suspend fun fetchClientSecret(): String? {
-            return embeddedComponentManager.fetchClientSecret()
+            return componentCoordinator.fetchClientSecret()
         }
 
         override fun getInitialParams(context: Context): ConnectInstanceJs {
-            return embeddedComponentManager.getInitialParams(context)
+            return componentCoordinator.getInitialParams(context)
         }
 
         override suspend fun onPermissionRequest(activity: Activity, request: PermissionRequest) {
@@ -286,7 +288,7 @@ internal class StripeConnectWebViewContainerViewModel(
             }
 
             // all calls to PermissionRequest must be on the main thread
-            val isGranted = embeddedComponentManager.requestCameraPermission(activity) ?: return
+            val isGranted = componentCoordinator.requestCameraPermission(activity) ?: return
             withContext(Dispatchers.Main) {
                 if (isGranted) {
                     logger.debug("($loggerTag) Granting permission - user accepted permission")
@@ -305,7 +307,7 @@ internal class StripeConnectWebViewContainerViewModel(
         ) {
             var result: Array<Uri>? = null
             try {
-                result = embeddedComponentManager.chooseFile(activity, requestIntent)
+                result = componentCoordinator.chooseFile(activity, requestIntent)
             } finally {
                 // Ensure `filePathCallback` always gets a value.
                 filePathCallback.onReceiveValue(result)
@@ -316,7 +318,7 @@ internal class StripeConnectWebViewContainerViewModel(
             activity: Activity,
             message: OpenFinancialConnectionsMessage,
         ) {
-            val result = embeddedComponentManager.presentFinancialConnections(
+            val result = componentCoordinator.presentFinancialConnections(
                 activity = activity,
                 clientSecret = message.clientSecret,
                 connectedAccountId = message.connectedAccountId,
@@ -328,7 +330,7 @@ internal class StripeConnectWebViewContainerViewModel(
         }
 
         override fun onReceivedPageDidLoad(pageViewId: String) {
-            webView.updateConnectInstance(embeddedComponentManager.appearanceFlow.value)
+            webView.updateConnectInstance(componentCoordinator.appearanceFlow.value)
             updateState { copy(pageViewId = pageViewId) }
 
             // right now view onAttach and begin load happen at the same time,
@@ -407,7 +409,7 @@ internal class StripeConnectWebViewContainerViewModel(
         val stripeIntentLauncher: StripeIntentLauncher = StripeIntentLauncherImpl(),
         val logger: Logger = Logger.getInstance(enableLogging = BuildConfig.DEBUG),
     ) {
-        val analyticsService = embeddedComponentManager.getComponentAnalyticsService(embeddedComponent)
+        val analyticsService = embeddedComponentManager.coordinator.getComponentAnalyticsService(embeddedComponent)
     }
 }
